requirements.txt
# Core Dependencies
RPi.GPIO>=0.7.1
smbus2>=0.4.2

# Sensor Libraries
adafruit-circuitpython-max30105>=2.1.0
adafruit-blinka>=8.20.0
w1thermsensor>=2.0.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0

# Web Dashboard
Flask>=2.3.0
Flask-CORS>=4.0.0

# Alert System
twilio>=8.5.0

# Configuration
PyYAML>=6.0

# Database
# SQLite3 is included with Python

# Utilities
python-dateutil>=2.8.2

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation (optional)
Sphinx>=7.0.0

# Logging
colorlog>=6.7.0
